pipeline/requirements.txt
fastapi==0.65.1
numpy==1.19.5
pandas==1.2.3
scikit-learn==0.24.2
tensorflow-serving-api==2.4.0
tensorflow==2.4.0
uvicorn==0.13.4